fixed some bugs

+++ src/main/java/de/thkoeln/ra/team3/risc_thk_simulator/guiTemplates/PC.java
@@ -9,7 +9,7 @@ public class PC {
 	private Label str;
 	private Label pc;
 	
-	public PC(int pc) {
+	public PC(long pc) {
 		this.str = new Label("PC:");
 		this.pc = new Label(""+pc);
